fix: rename daoQueries->dao-queries
