+++ src/components/daos/dao-chat-modal.tsx
@@ -19,7 +19,7 @@ import AgentWalletSelector from "../chat/agent-selector";
 import { CreateThreadButton } from "@/components/threads/CreateThreadButton";
 import { useChatStore } from "@/store/chat";
 import { useSessionStore } from "@/store/session";
-import { fetchDAOExtensions, fetchToken } from "@/queries/daoQueries";
+import { fetchDAOExtensions, fetchToken } from "@/queries/dao-queries";
 import type { DAO, Token, Extension } from "@/types/supabase";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import ThreadList from "@/components/threads/thread-list";
