+++ src/components/profile/agent-prompt-form.tsx
@@ -28,7 +28,7 @@ import {
 import { Badge } from "@/components/ui/badge";
 import { WalletInfoCard } from "./agent-wallet-info";
 
-import { fetchDAOs } from "@/queries/daoQueries";
+import { fetchDAOs } from "@/queries/dao-queries";
 import { fetchAgents } from "@/queries/agent-queries";
 import {
   fetchAgentPrompts,
